+++ corbeans-spring-boot-bnms-starter/src/main/kotlin/com/github/manosbatsis/corbeans/spring/boot/corda/bnms/message/MembershipRequestMessage.kt
@@ -19,14 +19,23 @@
  */
 package com.github.manosbatsis.corbeans.spring.boot.corda.bnms.message
 
+import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize
+import com.fasterxml.jackson.databind.deser.std.JsonNodeDeserializer
 import io.swagger.annotations.ApiModel
 import io.swagger.annotations.ApiModelProperty
 import net.corda.core.serialization.CordaSerializable
-import javax.json.JsonObject
 
+/**
+ * A flexible membership request message that allows arbitrary membership metadata to be used
+ * in combination with a custom
+ */
 @CordaSerializable
 @ApiModel(description = "A message with the information necessary to create or ammend a membership request.")
-open class MembershipRequestMessage(
-        @ApiModelProperty(value = "The BNO party name") var party: String,
-        @ApiModelProperty(value = "The membership metadata") var membershipMetadata: JsonObject
+class MembershipRequestMessage(
+        @ApiModelProperty(value = "The BNO party name")
+        val party: String,
+        @ApiModelProperty(value = "The membership metadata")
+        @JsonDeserialize(using = JsonNodeDeserializer::class)
+        var membershipMetadata: JsonNode? = null
 )